requirements.txt
akshare>=1.13.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
pyarrow>=15.0.0
fastparquet>=2024.2.0
google-generativeai>=0.3.0
